Refactor: Extract logging to independent module
