+++ FIX_SUMMARY.txt
@@ -1,0 +1,69 @@
+================================================================================
+CIRCULAR IMPORT FIX - SUMMARY
+================================================================================
+
+PROBLEM IDENTIFIED:
+-------------------
+ImportError: cannot import name 'MonitorController' from partially initialized 
+module 'monitor_controller' (most likely due to a circular import)
+
+Circular dependency chain:
+  monitornap.py → monitor_controller.py → monitornap.py (CIRCULAR!)
+
+
+SOLUTION IMPLEMENTED:
+---------------------
+Created a new independent logging module to break the circular dependency.
+
+BEFORE (Circular Dependency):
+-----------------------------
+  monitornap.py
+    ├── imports: monitor_controller, ui_components
+    └── contains: log_message() function
+           ↑
+           │ (imports log_message)
+           │
+  monitor_controller.py
+    └── imports: monitornap.log_message  ← CIRCULAR!
+
+
+AFTER (Fixed):
+--------------
+  logging_utils.py (NEW!)
+    └── contains: log_message(), LOG_CACHE, DEBUG_MODE
+           ↑
+           ├─── monitornap.py
+           │      └── imports: monitor_controller, ui_components, logging_utils
+           │
+           ├─── monitor_controller.py
+           │      └── imports: logging_utils
+           │
+           └─── ui_components.py
+                  └── imports: logging_utils
+
+  ✓ NO CIRCULAR DEPENDENCIES
+
+
+FILES CHANGED:
+--------------
+[NEW]     logging_utils.py       - Independent logging module
+[MODIFIED] monitornap.py          - Import from logging_utils
+[MODIFIED] monitor_controller.py  - Import from logging_utils  
+[MODIFIED] ui_components.py       - Import from logging_utils
+
+
+VERIFICATION:
+-------------
+✓ All Python files compile without syntax errors
+✓ No circular import dependencies detected
+✓ Import structure validated
+✓ Ready for PyInstaller packaging
+
+
+NEXT STEPS:
+-----------
+1. Test the application in development mode
+2. Rebuild with PyInstaller
+3. Verify the packaged executable runs without import errors
+
+================================================================================
